[IMP] move generate data in wizard
