+++ report/report_account_standard_report.xml
@@ -7,33 +7,33 @@
             <t t-set="data_report_dpi" t-value="110"/>
             <t t-call="report.internal_layout">
                 <div class="page" style="font-size:11px">
-                    <h2 t-esc="data['form']['name_report']"/>
+                    <h2 t-esc="data['name_report']"/>
                     <div class="row">
                         <div class="col-xs-3">
                             <strong>Company:</strong>
                             <p t-esc="res_company.name"/>
                         </div>
                         <div class="col-xs-3">
-                            <t t-if="data['form']['date_from']">
+                            <t t-if="data['date_from']">
                                 <strong>Start Date :</strong>
-                                <span t-esc="data['form']['date_from']"/><br/></t>
-                            <t t-if="data['form']['date_to']">
+                                <span t-esc="data['date_from']"/><br/></t>
+                            <t t-if="data['date_to']">
                                 <strong>End Date :</strong>
-                                <span t-esc="data['form']['date_to']"/></t>
-                            <p t-if="data['form']['with_init_balance']">With Initial Balance.</p>
+                                <span t-esc="data['date_to']"/></t>
+                            <p t-if="data['with_init_balance']">With Initial Balance.</p>
                         </div>
                         <div class="col-xs-3">
                             <strong>Target Moves:</strong>
-                            <p t-if="data['form']['target_move'] == 'all'">All Entries</p>
-                            <p t-if="data['form']['target_move'] == 'posted'">All Posted Entries</p>
+                            <p t-if="data['target_move'] == 'all'">All Entries</p>
+                            <p t-if="data['target_move'] == 'posted'">All Posted Entries</p>
                         </div>
                         <div class="col-xs-3">
-                            <strong t-if="data['form']['reconciled']">With Reconciled Entries</strong>
-                            <strong t-if="data['form']['reconciled'] == False">Only UnReconciled Entries</strong>
-                            <p t-if="data['form']['rem_futur_reconciled']">With entries matched with other entries dated after End Date.</p>
+                            <strong t-if="data['reconciled']">With Reconciled Entries</strong>
+                            <strong t-if="data['reconciled'] == False">Only UnReconciled Entries</strong>
+                            <p t-if="data['rem_futur_reconciled']">With entries matched with other entries dated after End Date.</p>
                         </div>
                     </div>
-                    <table class="table table-condensed" t-if="data['form']['type_ledger'] == 'partner'">
+                    <table class="table table-condensed" t-if="data['type_ledger'] == 'partner'">
                         <thead>
                             <tr>
                                 <th class="col-xs-1">Code</th>
@@ -66,7 +66,7 @@
                         </t>
                     </table>
 
-                    <table class="table table-condensed" t-if="data['form']['type_ledger'] == 'open'">
+                    <table class="table table-condensed" t-if="data['type_ledger'] == 'open'">
                         <thead>
                             <tr>
                                 <th class="col-xs-1">Total</th>
@@ -101,7 +101,7 @@
                                 <th class="col-xs-1">Account</th>
                                 <th class="col-xs-1">Journal entries</th>
                                 <th class="col-xs-3">Ref</th>
-                                <th class="col-xs-3" t-if="data['form']['type_ledger'] != 'partner'">Partner</th>
+                                <th class="col-xs-3" t-if="data['type_ledger'] != 'partner'">Partner</th>
                                 <th class="col-xs-1">Due Date</th>
                                 <th class="col-xs-1 text-right">Debit</th>
                                 <th class="col-xs-1 text-right">Credit</th>
@@ -119,20 +119,20 @@
                                     <td colspan="4">
                                         <strong t-esc="group_by_top(data, o, 'name')"/>
                                     </td>
-                                    <td t-if="data['form']['type_ledger'] != 'partner'"></td>
+                                    <td t-if="data['type_ledger'] != 'partner'"></td>
                                     <td class="text-right">
-                                        <strong t-if="data['form']['sum_group_by_top'] and sum_group_by(data, o, 's_debit')" t-esc="sum_group_by(data, o, 'debit')" t-esc-options='{"widget": "monetary", "display_currency": res_company.currency_id}'/>
+                                        <strong t-if="data['sum_group_by_top'] and sum_group_by(data, o, 's_debit')" t-esc="sum_group_by(data, o, 'debit')" t-esc-options='{"widget": "monetary", "display_currency": res_company.currency_id}'/>
                                     </td>
                                     <td class="text-right">
-                                        <strong t-if="data['form']['sum_group_by_top'] and sum_group_by(data, o, 's_credit')" t-esc="sum_group_by(data, o, 'credit')" t-esc-options='{"widget": "monetary", "display_currency": res_company.currency_id}'/>
+                                        <strong t-if="data['sum_group_by_top'] and sum_group_by(data, o, 's_credit')" t-esc="sum_group_by(data, o, 'credit')" t-esc-options='{"widget": "monetary", "display_currency": res_company.currency_id}'/>
                                     </td>
                                     <td class="text-right" style="white-space:nowrap">
-                                        <strong t-if="data['form']['sum_group_by_top']" t-esc="sum_group_by(data, o, 'debit - credit')" t-esc-options='{"widget": "monetary", "display_currency": res_company.currency_id}'/>
+                                        <strong t-if="data['sum_group_by_top']" t-esc="sum_group_by(data, o, 'debit - credit')" t-esc-options='{"widget": "monetary", "display_currency": res_company.currency_id}'/>
                                     </td>
                                     <td t-if="amount_currency"></td>
                                     <td></td>
                                 </tr>
-                                <tr t-foreach="lines(data, o)" t-as="line" t-if="data['form']['summary'] == False">
+                                <tr t-foreach="lines(data, o)" t-as="line" t-if="data['summary'] == False">
                                     <td style="white-space:nowrap">
                                         <i t-if="line['type_line'] == 'init'" t-esc="line['date']"></i>
                                         <strong t-if="line['type_line'] == 'total'" t-esc="line['date']"></strong>
@@ -154,7 +154,7 @@
                                         <i t-if="line['type_line'] == 'init'" t-esc="line['displayed_name']"></i>
                                         <span t-if="line['type_line'] == 'normal'" t-esc="line['displayed_name']"></span>
                                     </td>
-                                    <td t-if="data['form']['type_ledger'] != 'partner'">
+                                    <td t-if="data['type_ledger'] != 'partner'">
                                         <i t-if="line['type_line'] == 'init'" t-esc="line['partner_name']"></i>
                                         <span t-if="line['type_line'] == 'normal'" t-esc="line['partner_name']"></span>
                                     </td>
