+++ wizard/account_standard_report_view.xml
@@ -7,7 +7,7 @@
             <field name="arch" type="xml">
                 <form string="Report Options" class="oe_form_configuration">
                     <header>
-                        <button name="check_report" string="Print" type="object" default_focus="1" class="oe_highlight"/>
+                        <button name="print_pdf_report" string="Print" type="object" default_focus="1" class="oe_highlight"/>
                         <button string="Cancel" class="oe_link" special="cancel"/>
                     </header>
                     <field name="company_id" invisible="1"/>
